Implement tab content components for TeamAnalysisResultPage
