+++ frontend/src/components/analysis/TabContent/index.ts
@@ -2,3 +2,4 @@ export { default as SummaryTab } from './SummaryTab'
 export { default as TopicsTab } from './TopicsTab'
 export { default as ContributorsTab } from './ContributorsTab'
 export { default as HighlightsTab } from './HighlightsTab'
+export type { TabContentProps } from './types'
