+++ frontend/src/components/analysis/TabContent/types.ts
@@ -1,0 +1,13 @@
+import { FC } from 'react'
+
+export interface TabContentProps {
+  analysis: any
+  channel: any
+  processedAnalysis: any
+  reportResult: Record<string, unknown> | null
+  isTeamAnalysis: any
+  customStyles: any
+  workspaceUuid?: string
+}
+
+export type TabContentComponent = FC<TabContentProps>
