+++ frontend/src/components/common/index.ts
@@ -1,1 +1,2 @@
 export { default as ErrorBoundary } from './ErrorBoundary'
+export { default as ChannelAnalysisList } from '../integration/ChannelAnalysisList'
