+++ frontend/src/pages/integration/TeamAnalysisResultPage.tsx
@@ -1,4 +1,4 @@
-import React, { useState, FC } from 'react'
+import React, { useState } from 'react'
 import {
   Badge,
   Box,
@@ -6,9 +6,6 @@ import {
   BreadcrumbItem,
   BreadcrumbLink,
   Button,
-  Card,
-  CardBody,
-  CardHeader,
   Center,
   Flex,
   Grid,
@@ -46,147 +43,9 @@ import {
 import { Link, useNavigate } from 'react-router-dom'
 import ErrorBoundary from '../../components/common/ErrorBoundary'
 import { useAnalysisData } from '../../hooks'
-import { ServiceResource } from '../../lib/integrationService'
-import { renderPlainText, extractSectionContent, isObviouslyNotJson } from '../../utils/textRenderer'
+import { extractSectionContent, isObviouslyNotJson } from '../../utils/textRenderer'
+import { SummaryTab, TopicsTab, ContributorsTab, HighlightsTab } from '../../components/analysis/TabContent'
 
-interface ChannelAnalysisListProps {
-  title: string
-  reportResult: Record<string, unknown> | null
-  currentResources: ServiceResource[]
-  integrationId: string
-  filterFn?: (analysis: Record<string, unknown>) => boolean
-  contentField: string
-  emptyMessage?: string
-  workspaceUuid?: string
-}
-
-/**
- * A reusable component for displaying individual channel analyses in the team analysis view
- */
-const ChannelAnalysisList: FC<ChannelAnalysisListProps> = ({
-  title,
-  reportResult,
-  currentResources,
-  integrationId,
-  filterFn = () => true,
-  contentField,
-  emptyMessage = 'No information available.',
-}) => {
-  const navigate = useNavigate()
-
-  const filteredAnalyses =
-    reportResult?.resource_analyses &&
-    Array.isArray(reportResult.resource_analyses)
-      ? reportResult.resource_analyses.filter(filterFn)
-      : []
-
-  if (
-    !reportResult ||
-    !reportResult.resource_analyses ||
-    !Array.isArray(reportResult.resource_analyses) ||
-    reportResult.resource_analyses.length === 0
-  ) {
-    return null
-  }
-
-  return (
-    <>
-      <Heading size="md" mb={3} mt={6}>
-        {title}
-      </Heading>
-      {filteredAnalyses.length > 0 ? (
-        filteredAnalyses.map((channelAnalysis) => (
-          <Card
-            key={channelAnalysis.id}
-            variant="outline"
-            mb={4}
-            boxShadow="sm"
-          >
-            <CardHeader pb={1}>
-              <Flex justify="space-between" align="center">
-                <Heading size="sm" color="purple.600">
-                  {/* Try to get channel name from different sources with fallbacks */}
-                  {channelAnalysis.results?.channel_name ||
-                    channelAnalysis.channel_name ||
-                    channelAnalysis.resource_name ||
-                    currentResources.find(
-                      (r) => r.id === channelAnalysis.resource_id
-                    )?.name ||
-                    `Channel ${channelAnalysis.resource_id.substring(0, 8)}...`}
-                </Heading>
-                <Badge
-                  colorScheme={
-                    channelAnalysis.status === 'COMPLETED'
-                      ? 'green'
-                      : channelAnalysis.status === 'FAILED'
-                        ? 'red'
-                        : 'yellow'
-                  }
-                >
-                  {channelAnalysis.status === 'PENDING' ? (
-                    <HStack spacing={1}>
-                      <Text>PENDING</Text>
-                      <Spinner size="xs" />
-                    </HStack>
-                  ) : (
-                    channelAnalysis.status
-                  )}
-                </Badge>
-              </Flex>
-            </CardHeader>
-            <CardBody pt={2}>
-              {channelAnalysis[contentField] ? (
-                <Box className="analysis-content">
-                  {/* Extract channel-specific integration_id if available */}
-                  {renderPlainText(
-                    typeof channelAnalysis[contentField] === 'string'
-                      ? channelAnalysis[contentField].replace(
-                          /(\d+\.\s)/g,
-                          '\n$1'
-                        )
-                      : 'No content available',
-                    channelAnalysis.workspace_uuid
-                  )}
-                </Box>
-              ) : (
-                <Text fontSize="sm" color="gray.500">
-                  {emptyMessage}
-                </Text>
-              )}
-              {channelAnalysis.status === 'COMPLETED' && (
-                <Button
-                  size="sm"
-                  colorScheme="purple"
-                  variant="outline"
-                  mt={3}
-                  onClick={() => {
-                    const resourceId = String(channelAnalysis.resource_id)
-                    const analysisId = String(channelAnalysis.id)
-                    console.log('Channel analysis debug:', {
-                      resourceId,
-                      analysisId,
-                    })
-
-                    navigate(
-                      `/dashboard/integrations/${integrationId}/channels/${resourceId}/analysis/${analysisId}`
-                    )
-                  }}
-                >
-                  View Full Analysis
-                </Button>
-              )}
-            </CardBody>
-          </Card>
-        ))
-      ) : (
-        <Text fontSize="sm" color="gray.500" mt={4}>
-          No completed channel analyses with {contentField.replace(/_/g, ' ')}{' '}
-          information available.
-        </Text>
-      )}
-    </>
-  )
-}
 
 /**
  * Improved Analysis Result Page component that provides a more accessible
@@ -862,120 +721,54 @@ Generated using Toban Contribution Viewer with ${modelUsed}
           <TabPanels>
             {/* Summary Tab */}
             <TabPanel sx={customStyles.tabPanel}>
-              <Card variant="outline">
-                <CardBody>
-                  {processedAnalysis?.fixedChannelSummary
-                    ? renderPlainText(
-                        processedAnalysis.fixedChannelSummary,
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )
-                    : renderPlainText(
-                        analysis?.channel_summary || 'No summary available',
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )}
-                </CardBody>
-              </Card>
-
-              {/* For team analysis, show individual channel summaries */}
-              {isTeamAnalysis && reportResult && (
-                <ChannelAnalysisList
-                  title="Channel Summaries"
-                  reportResult={reportResult}
-                  currentResources={[]}
-                  integrationId=""
-                  contentField="resource_summary"
-                  emptyMessage="No channel summaries available."
-                />
-              )}
+              <SummaryTab
+                analysis={analysis}
+                channel={channel}
+                processedAnalysis={processedAnalysis}
+                reportResult={reportResult}
+                isTeamAnalysis={isTeamAnalysis}
+                customStyles={customStyles}
+                workspaceUuid={channel?.metadata?.workspace_uuid as string}
+              />
             </TabPanel>
 
             {/* Topics Tab */}
             <TabPanel sx={customStyles.tabPanel}>
-              <Card variant="outline">
-                <CardBody>
-                  {processedAnalysis?.fixedTopicAnalysis
-                    ? renderPlainText(
-                        processedAnalysis.fixedTopicAnalysis,
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )
-                    : renderPlainText(
-                        analysis?.topic_analysis ||
-                          'No topic analysis available',
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )}
-                </CardBody>
-              </Card>
-
-              {/* For team analysis, show individual channel topic analyses */}
-              {isTeamAnalysis && reportResult && (
-                <ChannelAnalysisList
-                  title="Channel Topic Analyses"
-                  reportResult={reportResult}
-                  currentResources={[]}
-                  integrationId=""
-                  contentField="topic_analysis"
-                  emptyMessage="No topic analyses available for individual channels."
-                />
-              )}
+              <TopicsTab
+                analysis={analysis}
+                channel={channel}
+                processedAnalysis={processedAnalysis}
+                reportResult={reportResult}
+                isTeamAnalysis={isTeamAnalysis}
+                customStyles={customStyles}
+                workspaceUuid={channel?.metadata?.workspace_uuid as string}
+              />
             </TabPanel>
 
             {/* Contributors Tab */}
             <TabPanel sx={customStyles.tabPanel}>
-              <Card variant="outline">
-                <CardBody>
-                  {processedAnalysis?.fixedContributorInsights
-                    ? renderPlainText(
-                        processedAnalysis.fixedContributorInsights,
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )
-                    : renderPlainText(
-                        analysis?.contributor_insights ||
-                          'No contributor insights available',
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )}
-                </CardBody>
-              </Card>
-
-              {/* For team analysis, show individual channel contributor insights */}
-              {isTeamAnalysis && reportResult && (
-                <ChannelAnalysisList
-                  title="Channel Contributor Insights"
-                  reportResult={reportResult}
-                  currentResources={[]}
-                  integrationId=""
-                  contentField="contributor_insights"
-                  emptyMessage="No contributor insights available for individual channels."
-                />
-              )}
+              <ContributorsTab
+                analysis={analysis}
+                channel={channel}
+                processedAnalysis={processedAnalysis}
+                reportResult={reportResult}
+                isTeamAnalysis={isTeamAnalysis}
+                customStyles={customStyles}
+                workspaceUuid={channel?.metadata?.workspace_uuid as string}
+              />
             </TabPanel>
 
             {/* Highlights Tab */}
             <TabPanel sx={customStyles.tabPanel}>
-              <Card variant="outline">
-                <CardBody>
-                  {processedAnalysis?.fixedKeyHighlights
-                    ? renderPlainText(
-                        processedAnalysis.fixedKeyHighlights,
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )
-                    : renderPlainText(
-                        analysis?.key_highlights || 'No highlights available',
-                        String(channel?.metadata?.workspace_uuid || '')
-                      )}
-                </CardBody>
-              </Card>
-
-              {/* For team analysis, show individual channel highlights */}
-              {isTeamAnalysis && reportResult && (
-                <ChannelAnalysisList
-                  title="Channel Highlights"
-                  reportResult={reportResult}
-                  currentResources={[]}
-                  integrationId=""
-                  contentField="key_highlights"
-                  emptyMessage="No highlights available for individual channels."
-                />
-              )}
+              <HighlightsTab
+                analysis={analysis}
+                channel={channel}
+                processedAnalysis={processedAnalysis}
+                reportResult={reportResult}
+                isTeamAnalysis={isTeamAnalysis}
+                customStyles={customStyles}
+                workspaceUuid={channel?.metadata?.workspace_uuid as string}
+              />
             </TabPanel>
           </TabPanels>
         </Tabs>
